Reorganize Board class
