+++ lib/board_display.rb
@@ -1,0 +1,41 @@
+module BoardDisplay
+  def show_board
+    puts ''
+    grid.each_with_index do |row, row_index|
+      numbers_column(row_index)
+      row_index.even? ? white_black_row(row) : black_white_row(row)
+    end
+    letter_coordinates
+  end
+
+  def white_black_row(row)
+    row.each_with_index do |piece, column|
+      print column.even? ? white_square(piece) : black_square(piece)
+    end
+    print "\n"
+  end
+  
+  def black_white_row(row)
+    row.each_with_index do |piece, column|
+      print column.even? ? black_square(piece) : white_square(piece)
+    end
+    print "\n"
+  end
+
+  def white_square(piece)
+    piece.nil? ? "\e[48;5;245m   \e[0m" : "\e[48;5;245m #{piece.symbol} \e[0m"
+  end
+
+  def black_square(piece)
+    piece.nil? ? "\e[48;5;237m   \e[0m" : "\e[48;5;237m #{piece.symbol} \e[0m"
+  end
+
+  def letter_coordinates
+    puts "     #{('a'..'h').to_a.join('  ')}"
+  end
+
+  def numbers_column(index)
+    numbers = (1..8).to_a.reverse
+    print "  #{numbers[index]} "
+  end
+end
